rename: read-dinnerPlans -> read-dinner-plans
